Impose dependency requirements in CMake as described in README
For ROOT, HepMC and Rivet the versions requirements have been set in a
less standard way and namely not simply using the version argument of
find_package, but rather manually testing the version in case the
dependency was found. The advantage in this case is that more
informative output can be given to the user, while CMake would silently
skip the found version if the version requirement is not fulfilled.

To be fully precise, CMake would in general report if an
older-than-requested version is found, but this requires to use
find_package without the QUIET option and in many cases, then, the
CMake output blows up because of other messages which we do not want.