Properly find Doxygen in CMake code using modern portable approach
The way Doxygen was located was broken on macOS in case the user
installed Doxygen using a .dmg package. Moreover Doxygen offers a
module to treat Doxygen as package and this has been now used. To build
documentation one could use the CMake doxygen_add_docs function, but
given that we have further operations to do in order to build the user
documentation, I refrained from using it.