Remove all occurrences of speed of light in units of measure
SMASH works in natural units and hence it is confusing to still specify
c in units of quantities like writing fm/c for time or GeV/c for
momenta. I took the possibility to also fix some wrong units (fm/c was
given as length unit) and to fix as well the clock description which
was inaccurate.