Implement more robust behavior about requested output formats
An empty "Format" list is not accepted anymore and "None" can now be
(alone) specified to suppress a content output. Invalid formats now
trigger a SMASH termination, too. The user has to be aware that no
output is going to be produced and, if this is their need, request so.