Manually implement throwing destructor and move operations
The Configuration class must now be fully consumed before being
destructed. This might be see as very strict and not too
developer-friendly, but it makes it harder to misuse the class. In
particular, a take call with a typo in the key labels and a default
value would silently make the default value to always being used. This
change at least would make the code abort if the configuration file
contains that wrongly taken key. The fact that the destructor can now
throw (implemented in a safe way w.r.t. possible stack unwiding)
requires move operations to be manually implemented, because we need
the moved from object to result cleared up at destruction, what the
default compiler-provided move operations would not guarantee.

This change breaks most of tests, i.e. every that is violating the
enforced behaviour. Next I will add a clear() method to the
Configuration class in order to easily be able to clean up instances in
tests cases where it is not really relevant to completely consume all
keys. Note that this new implemented behavior might also make
sub-optimal usage of the class emerge and all failures should be
carefully analysed.