Build image even with unsuccessful SMASH compilation
We decided that the Dockerfiles of the development branch should focus only on the SMASH development version. So, in case of breaking changes in the dependencies (e.g. here we removed the boost library) the stable public version of SMASH might not compile anymore. With this commit, if SMASH does not compile, the copy of the public repository is deleted and the rest of the image is built. In addition, the base Dockerfile now uses the syntax specifier, so exploit BuildKit and allow to use heredocs. Unrelated to these changes, we added a couple of general purpose useful programs which were left out.