Improve Configuration class implementation
It has been noticed that methods to descend the YAML tree looking for a
not existing key were not changing the object in a visible way, but were
allocating memory underneath due to the behaviour of the YAML::Node
non-const operator[] method. This is undesired and potentially leading
to large memory usage if repetitevely look for not-existing keys
(improbable but still possible). The substcript operator is now not used
at all when an expected-to-exist node is looked for in the tree, while
it is used when a node has to be located and possibly created.