Fix bug in Configuration::set_value making it work on empty instances
This has been achieved using an indirect approach. The reason why the
method was not working with empty instances was that the YAML::Node
root_node_ member was initialised to a null node (in the YAML sense)
when the object was created as empty. However, such a member in the
SMASH case is always meant to be a map and this has been enforced
now. New tests to check the proper behaviour have been added.