Move list of Doxygen input files to CMakeLists file
Having the Doxygen input file hard-coded in the Doxyfile.in and
DoxyUser.in files leads not only to duplication but requires changes to
stay in sync between the two files and this is error-prone. Actually the
two files were already out of sync. Since some tags of the *.in files
were already configured by the configure_file CMake function, this was
also used to set the list of input files to be used. An index file to
fix documentation pages order has already be included in the list of
files, although the file is still missing and will created next.