Create index page for documentation using anchor prefix
An index page has been created to fix ordering of documentation pages.
This has implied that all \page commands in the codebase using two
arguments have been replaced by the same using one argument only, so
that the documentation titles are now all in the index file. Since
anchor names were in some cases difficult to search for, a "doxypage_"
prefix has been introduced in order to then rename them later to have in
the end a consistent style.