Re-enable default use of fused mathematical expressions in smash only
In the end, after more thinking, it has been decided to allow compilers
to use fused mathematical expressions, since this is likely to be
available in HPC machines and probably leads to a minor speed-up.
However, tests should be run with this feature turned off and this has
been imposed in CMake code.