Fix bug in grid constructor which was incomplete in box modus
In corner cases, when the calculated number of cells in some direction
resulted to be 0, this was set to 1 but this is not valid in the box
modus. In the box modus an error should be given, because the user has to
choose the time step and the size of the box to ensure that twice the
minimum cell size fits into the box (at least). The error was
implemented, but it was triggered only when the calculated number of
cells was 1 and not 0.