Encapsulate Clebsch-Gordan lookup table in new class
Both the lookup table and the method to retrieve or calculate the CG
coefficients have been put in a new class, hiding as much as possible
from client code. The class interface only offers a static method to
obtain a coefficient and a struct to keep together three spins
information. Tests about the pure calculation of coefficients have been
separated from those of functions used aroung in SMASH (which internally
needs CG coefficients).