Implement lookup mechanism for Clebsch-Gordan coefficients
Values of the Clebsch-Gordan coefficients up to isospin 3/2 for the two
incoming particles have been tabulated. Only non-zero coefficients have
been stored and a mechanism to calculate any non-tabulated coefficient
has been implemented. At the moment, the lookup is done in the
smash::clebsch_gordan namespace, but as global variable. This is not nice
and it is only a temporary implementation. It should be feasible to wrap
the lookup table in a ClebschGordan class, which is meant to replace the
clebsch_gordan namespace. The at the moment existing free functions to
calculate and look up a coefficient should be made members of that class,
probably exposing only one generic (static) method to get a coefficient
(it does not matter from the client perspective whether the coefficient
is calculated or looked up).