Improve jet implementation in sphere mode
The input key to specify the jet PDG code was said to be optional in the
documentation but, indeed, it was rather meant to be mandatory when the
section is provided. This is now reflected in the code and, if the jet
section is provided without the PDG code, an error is raised.