Fix failing creation of automatically determined lattice
In the refactoring of the input keys, the behaviour of having an empty
Lattice section to request an automatically determined lattice has been
broken. A new "Automatic" key has been introduced to explicitly request
to fully automatically create the lattice. If this new key is missing, at
least one among "Cell_Number", "Origin" and "Sizes" need to be specified.