Improve code design about how to validate particle lists
Many more or less minor changes have been done to the existing code to
add/remove particles. In particular, the new free function to create a
valid SMASH particle matching provided quantities (used now both in
ListModus and in Experiment) is returning a ParticleData instead of
using the 4-momentum as in-out parameter. Furthermore, a new free
function to compare two ParticleData instances at a given time has been
implemented in order to make the new code in Experiment easier to
follow. Finally, the search for particles to be removed to ensure that
they do exist in SMASH is now done using a std::find_if algorithm.