Fix couple of bugs in validation of add/remove lists
The particle being validated had to be adjusted in place but it was
accessed through an iterator and first copied (instead of being referred)
into a variable. This was implying that the particle was not adjusted at
all. In turn, this bug was hiding another one. The particle adjustment
was done potentially changing id/process of the particle instead of its
4-momentum only. This was making the FreeforallAction being discarded as
the particles looked invalid.

The std::vector<>::erase method was also not properly used and this was
fixed, too.