Add PdgCode constructor for hexadecimal nuclei codes
Before this change, it was not possible to create PDG codes for nuclei
using a hexadecimal code, as a 10 digits long code would have exceeded
the int32_t parameter maximum. As consequence, nuclei codes are declared
as compile-time constants as done for other particles (before it was not
possible).

Note that at the moment the documentation is somehow broken, because
Doxygen does not recognize the new constructor template, which is
breaking the class documentation.