Refine logic about lattice creation improving implementation
The meaning of the "Automatic" key about lattice creation has now changed
from fully automatic to simply automatic. This means that if specified to
false, ALL geometric properties must be provided and, if specified to
true, at least one has to be omitted. This allowed to improve the
implementation as the automatic determination of the lattice geometry is
now skipped if the user does not request it.