Add work-around to fix parsing of input key
This ugly patch is now necessary to ship next release without this bug,
but a higher-level refactoring is needed and will be done in the next
months. In short, at the moment, the user input configuration is parsed
throughout the codebase and often input sections are extracted and handed
over to object constructors to parse them. In few cases, this is violated
and then it is harder to find working approaches. The way the input is
parsed is bad at many levels. To mention few, way too many SMASH classes
depend on Configuration functionality and key labels as well as default
values are spread all around in the code. It is planned to improve this
aspect.