Fix bug in reset function of UniformClock
The reset was done using the timestep_duration function that in some
cases (when close to the end time) returns the duration of the last
step instead of the output interval specified by the user. Since the
reset of the UniformClock class is meant to be exclusively used to reset
the output clock at the beginning of each event in Experiment, resetting
the clock at the beginning of each event should always do the same thing
and not produce different results depending on the amount of elapsed
time. In particular, in Experiment at the beginning of the events
successive to the first one, if the end time specified by the
user was not a multiple of the output interval, then the reset of the
output clock was leading to output times different to those of the first
event. Using the time-step duration stored in the clock at creation time
(specified by the user as Output_Interval) in the reset function solves
the issue.