Fix bug for End_Times equal or less than zero
If the End_Time is equal or less than zero, the code didn't run properly.
It started to run. Then it seemed to pause since nothing happened, but
didn't abort (the user was not prompted again). This fix forbids setting
the End_Time to a negative value or zero. It should only be a temporary
fix since the code should run for an End_Time greater than the initial time.