Modify introduction version of some keys from git history
With some auxiliary shell scripting, the oldest tag containing taking an
input key was found and inserted into SMASH. This has been done at the
moment only for those keys, whose depest level identifier is unique. In
the next upcoming commits it will be dealt with those keys that have the
same last level identifier but that can appear in different sections.
Notice that for some keys the search for the tag failed, as the key is
never explicitly taken (e.g. for the logger levels). These cases are few
will be dealt with manually.