Remove interface methods arguments name and make them throw
The class OutputInterface has been created with virtual methods of all
children but each child implements only some of them. The base class was
providing an empty implementation and this was dangerous in case a wrong
child not-implemented method was called. Now the base implementation
throws an exception. Morevoer, the documentation of parameters is done
in children and in the base class this has been removed, as there the
parameters name are irrelevant and would just trigger compiler warnings.