Revert OutputInterface implementation with throwing methods
It turned out that the Experiment class really uses unimplemented
methods of children classes of OutputInterface and relies on the fact
that the base class implementation is called. Hence assuming that the
base class methods should not be called if not implemented by children
was wrong. This design is pretty poor at many levels, but the code in
Experiment is too intricated to easily refactor it. The bottom line
contradiction is that on one side polymorphism seems wished to be used
and therefore having an abstract interface, but on the other side the
use of the pointers to the base class initialised with derived classes
in Experiment requires to know the type of the derived class, which
would speak against using polymorphism at all.