Treat potentials tests differently in locally hosted actions
In particular, the unit test has been estimated to fail roughy every
~200 events and we repeat it on failure up to three times. If it fails
three times in a row, the situation should be investigated, as the
probability to fail because of statistics is roughly 1.e-7 (namely
rwoughly zero). Furthermore, as the local machine hosting the action
might be busy and hence slower, the slowest physics test we have (i.e.
potentials_run) is run increasing the default ctest timeout. Please,
note that it is not possible to repeat such a test, as subsequent
identical runs would fail because of the previous "crash" due to the
timeout which would leave a smash.lock file in the output folder.