Set cpu type flag as parameter in Dockerfiles
Until now in the Dockerfiles to create Docker images we used the flag '-march=x86-64' to ensure the maximum compatibility on Intel/AMD architectures. However, one needs to change this flag for ARM64 cpus or to obtain a better performance on recent cpu models. With this commit we make easier to change the cpu type compilation flag when builiding the container image, leaving the old flag just as the default option.