Review install procedure completely
The install procedure now considers not only SMASH executable and
library but also all third-party libraries that are frozen into SMASH.
In particular, a "smash" folder gets created in the installation
directory specified to CMake and inside it the typical bin, lib, include
and share folders are created. All libraries (SMASH, einhard, Cuba and
YAML) are put together in the lib folder, while the header files in the
include folder are organised in sub-folders. The CMake module to find
SMASH should be now adjusted to correctly cope with the new folder tree.