Add low momentum approximation and test for EoM
The low momentum approximation is necessary when the momentum of the
considered particle is numerically very small, so that dividing by it
isn't possble.
The test compares the energy gradient needed for the equations of
motion with momentum-dependence to the gradient of the potential
expressed using the chain rule. This is done without momentum-
dependence, where both implementations should work.