Clean up and implement new feature of List* modi
The class was containing an unused member and was initializing the other
members in a suprising manner. I took the opportunity to make some
clean-up and to then implement the possibility to use a single file
without any indexing. In such a case, the member containing the file
prefix is actually storing the full filename and the function that
constructs the full filename checking if it exists was adjusted with
some simple logic. Also the mechanism to read from the following file
should occur if and only if there is potentially a following file.