Fix ListBox modus class constructor configuration parsing
The ListBox class was not working because it was reusing the parsing of
the configuration file of the parent List class, but the section in the
child case had a different name. This made a new difficulty arise: The
child class has to parse child-specific keys ('Length' in this case) but
the parent constructor was getting the configuration moved in input
Therefore, on one hand, it had to parse it all (non-empty configuration
throw when destroyed) and on the other hand the child could not retrieve
any more the child-specific key. In this commit the constructor of all
modi were changed to get the configuration by reference and not by value
but this is a large change, as all modi must have the same constructor
to allow Experiment to have a modus template parameter.