Implement new approach to find SMASH in a CMake project
When using SMASH as library, at the moment, the findSMASH.cmake module
is forcing the user to refer to the SMASH source code folder and its
build directory, instead of a safer installation directory. The CMake
module has been now extended to include this different approach. The
older one is still possible and which approach is used depends on which
environment variable is set. Refer to the documentation in the beginning
of the module for more information.

The module has been in general improved, too. In external projects it is
now possible to specify and request a SMASH version. Furthermore
existence of header folders is tested.