Add check on Experiment constructor path parameter
The output path passed to the Experiment class constructor was not
validated as e.g. done in the ExperimentBase::create function. If such a
constructor was used (for instance in an external software) and a not
existing path given (for example expecting that SMASH would have created
the output folder), a segmentation fault would have occurred trying to
create output files. Now the output path is validated and an error is
given, making explicit the assumption that such a path must be valid,
i.e. all directories in it must already exist.