Start implementing automatic creation od CITATION.cff file
At the moment only the authors are added and the remaining metadata are
missing. It has to be decided how to add them, but it might make sense
to add them to the .authors.json file which might then be renamed to a
different more general name (e.g. .metadata.json).