Fix bug when 'Power_Particle_Formation' key is used
The parsing of the 'String_Parameters' subsection was done extracting it
into a temporary config in the 'create_finder_parameters' function. At
that point all keys but the 'Power_Particle_Formation' were taken and
the remaining one was meant to be taken later from the Experiment
constructor. This of course led to a non-empty configuration be
destroyed and hence an exception been thrown. The single key of this
section is now taken in Experiment before creating the
ScatterActionFinder object, which in turn calls
create_finder_parameters.