Fix MathJax URL in documentation setup, add version to head
The previously used URL was not secure and this led to the wrong
visualization on GSI server. Pulling it from a secure server would have
been enough, but according to https://www.mathjax.org/cdn-shutting-down/
the previous URL has been dismissed and should not be used. I used a
not too recent fixed version 2.7.1 of MathJax, hence following the
recomendation not to always use the latest one to avoid surprises.

The codease version has been added to the header of all HTML pages.
That information was basically hidden to the user so far, but it is
probably important to have it well visible.