Use environment variable to let CMake find Eigen
This is better than explicitly setting the path when invoking CMake,
because the internal SMASH test to test the usage of SMASH as library
also uses CMake but does not get the correct Eigen location, if this is
not a standard one and is not set using environment variables. I took
this opportunity to improve the bash code always using braces when
expanding variables and invoking cmake as it should be done (.. at the
end of the command after options).