Add failing tests about CustomClock class
This class has many wrong aspects at the moment. Removing the times in
the past is not really removing them from the clock. Moreover the clock
can be ticked beyond its last time and access out of memory. I'd and
will throw an exception instead. The added tests fail because the access
out of memory now does not throw the expected exception.