Fix bug in CustomClock and make class safer to use
The method to remove times in the past was not really removing them, but
shifting them beyond the largest time in a valid but undefined
state. Now this is fixed. Furthermore, the methods to get the current
time and next time were potentially reading out of memory and have been
changed making them throw in such a case. The class documentation has
been completed, corrected and improved, too.