Make Default class private into Key object and adjust ctors
After all this is an implementation detail of the Key class and it does
not make any sense to expose it. Furthermore, by doing so, it is
possible to have a private constructor using it and the public interface
of the Key class gets more explicit and simplified in its implementation
because it is then possible to entirely delegate construction to the
private constructor.