Fix bug in accessing default value of required key
Due to the temporary duplication of the Initial_Conditions keys, the
ColliderModus class constructor has to deal with the possibility that
the user specifies that information in the configuration file in a
different way and therefore needs to set a default there. This was
previously set in the key, but I believe it is better to hard-code it in
the take method.