Introduce new Configuration::take method to take a Key
This is a first step towards a big refactoring at the end of which we
want the Configuration class to be able to take values via Key objects
and not hard-coded series of labels.

In this commit a new take method template has been introduced and the
full implementation has been delegated to a private method which is not
a template and which, then, can have its definition in the cc
file. Along the way it has been realised that, although C-style
strings (const char *) need to be used to interface with the YAML
library, there is not really any reason to stick to this type, rather
than using more modern C++ features. For example, std::string_view is
very well suited to pass around views on string-like objects. This type
of change is going to be part of next refactoring, too.