Change logic when taking a Key object from configuration
Now that the Configuration class has a method to take a key via a Key
object, this can be done with some better logic. Namely, it is possible
to automatically return the default valut if the taken key is not
present in the configuration. Of course, taking a key without a default
results in an error and to take a key whose default is dependent on
other parameters, a dedicated new method is introudced. This cannot be
called for keys which have a default value, though.

Note that the Configuration::take new method template returns a T and
not a Value. This does not require the client anymore to decide the
taken type on assignment and it leaves Configuration::Value as a truly
implementation detail. The drawbacks are that all taken keys cannot
always be assigned to strings (it was previously possible, but probably
not really nice to have) and that now the assignment of the taken value
to any variable undergoes the C++ conversion rules. It is in principle
possible to implicitly assign a double taken value to an integer loosing
precision, but this is not different than anywhere else.