Fix usage of GSL library about elliptic integrals
These functions were called passing a precision double variable as
second argument, which was set to 1.e-6. However, the library function
takes a gsl_mode_t as second argument, which is an unsigned int
variable. Therefore, the functions were called with 0 as mode which is
GSL_PREC_DOUBLE and which makes the library calculate the result up to
double precision (approximatively 2.e-16). Therefore the intention of
the code was not respected and I now put in code was it was really
happening.