Use KeyLabels type in some Configuration methods
This type is underneath a vector<std::strings> which is now used instead
of an initializer_list<const char *>. It might be argued that this is
less lightweight to be passed around, but all Configuration business is
in SMASH a one-time business done in the beginning of any simulation. It
is not a performance-relevant part and I prefer to use more expressive
code and hide lower level decisions. KeyLables must be a vector as we
use it elsewhere changing its size and introducing a new alias or type
here feels superfluous.