Implement mechanism to forbid taking same key twice
This has been achieved making the take method store in a private member
the labels of the taken key. Such operation is done only when taking an
existing key, and the labels are not stored if the key is not present in
the configuration and the method returns its default value.