Introduce new read methods which are analogous to take ones
These new method templates now accept a Key as parameter and do not
expose any more the low-level details of key labels or Value proxy
object. They have been implemented in a 1:1 way to the take methods
apart from the clear difference that read does not remove the Key from
the configuration and in this sense it is safe to be used several
times. Needless to say, the read methods are const, while take is not.