Add some utility methods to Key class
In particular, functionality for tests has been added. A Key object can
now return its labels in a YAML formatted way and also clone itself
dropping the top-level label. As a custom type trait was needed to
understand if a type was streamable, a dedicated file was added. This
might come in handy in the future.