Make Key class constructor arguments handier and descriptive
So far only std::initializer_list of labels were accepted, but this was
making it inconvenient (if not impossible) e.g. in tests to create keys
with labels already stored in a container different from the needed
one. I now decided to simply accept an std::vector<std::string> which
will be yes more expensive, but it is way clearer in a place that is
hardly performance critical (Key creation).