Fix logic on number of events depending on user input
The Experiment class was taking the Nevents key as optional with 0 as
default value and later on using has_value on it, which would have never
been true, since the key was taken. I refactored this logic making it
also easier to follow. The private member is not any more constant,
which is consistent with the other two integer about the minimum number
of events not being constant. Having constant members prevents move
operations and this is rarely used in C++.