Add InputKeys functionality to find some keys in database
These are needed in special cases when many similar keys are taken in a
look over a string (like for the logging area or the output format). In
that case client code does not know which key to take and this key will
be found using the here added methods.