Remove old interface for Configuration::take with default
This required a huge set of changes as expected. Most of them were
simply replacing the key labels with the appropriate InputKeys
member. However, some keys (Start_Time in box and sphere modi as well as
Fractional_Photons) were taken with a default, although not having one
and being stated as required in the documentation. This has been fixed
making them really mandatory.

Moreover, the take operations in the ListModus class constructor have
been refactored as a naive replacement was not really possible. Also the
orientation parsing logic in the Nucleus class has been simplified in
code, but left unchanged.

Note, that some tests are now failing as there is a mismatch in how the
taken keys are in the sub-configurations w.r.t. their labels. Now keys
are (almost) always taken with all labels, but different objects get
sub-configurations without the top-level labels. This will be fixed
later using the new method to extract a complete sub-configuration as
much as possible.