Introduce new lightweight database for input sections
This will enable to add a Configuration::has_section method which will
in turn be used when the now existing Configuration::has_value method
needs to check for the existence of a section and not of a scalar
key (in the YAML sense). The sections will now be used in the keys
database to avoid the section labels duplication. At the moment,
renaming a section would imply to potentially have to change lots of
key labels.