Add Configuration::has_section method and clean-up behaviour
The Configuration class now distinguishes between keys and
sections. Although from the YAML point of few a section is a key which
has a map value, we prefer to make a distinction in SMASH, i.e. we call
key any YAML key which is not a map (in the YAML sense) and we call
section keys which are YAML maps. This will make logic clearer from the
clients of the Configuration class.