Remove old interface for Configuration::has_value
This required modifications sparse everywhere in the codebase and I took
the opportunity to replace the old usage of the method distinguishing
between Configuration::has_section or the new has_value methods.

Since we want to take keys with full labels from the configuration
top-level, some additional logic has been needed mainly in the Nucleus
and its children classes. This is due to the fact that there are
identical keys for projectile and target and when a nucleus is
initialized it is totally irrelevant whether it is belonging to the
projectile or to the target. At the moment the conservative approach of
understanding from the configuration what is being parsed has been
chosen. This requires to check whether the projectile or target
subsection is present and then a reference to the key to be taken is set
before taking it.

Note that many tests still fail because the sub-configuration which are
being parsed around in the code are still not the correct ones. They
should be now extracted from the main configuration using the new
Configuration::extract_complete_sub_configuration method and so will be
done soon.