Use std::snprintf function instead of deprecated std::printf
Initially this was not done as the usage of std::sprintf was done in a
controlled context without real risks. However, it turned out that
recent compilers warn about this and this is rather annoying.
Furthermore it turned out from a quick benchmark that our usage does not
lead to any performance loss, as it was potentially possible.