Remove "Version" key and create doc page about removed keys
The "Version" key was deprecated and it has been removed now. This has
also been the opportunity to collect all removed keys in a dedicated
documentation page. Such a list will grow in the future and the needed
steps to remove keys should be straightforward. This has been included
in the documentation, too.