Answer to review comments.
Change `ColliderModus::fluid_lattice()` to return a const-ref, which required making const some methods in RectangularLattice.
Rename `IC_output_switch` to `IC_switch` in *experiment.h* because of the ongoing refactoring.
Also delete the run_initial_conditions_run test since a near-future PR will replace it, and it was failing for not including the Modi: Collider: Initial_Conditions section.