Remove empty, not overridden function from OutputInterface
Such a function was used in Experiment, but not in a consistent
way. Furthermore the function was declared as other similar ones in the
OutputInterface class, without being overridden in any of the children
classes. Probably this is just part of an old code change and it sneaked
into production.