Change OutputInterface methods to take an EventLabel
All output formats which do not average over multiple parallel ensemble
should print the ensemble number in the output file(s) or in its name.
At the moment this is not done and the using SMASH with multiple
parallel ensembles gives wrong number of events in most of output files.

Note that this commit breaks compilation and code will be fixed in
subsequent commits.