Fix default value of nuclei displacement input key
In commit e8f92f0 I used the new Configuration interface to take the
initial z-displacement key in the new way, i.e. use its default value if
the key was not given in the input file. Such a value was then divided
by 2. However, before the commit, the variable storing the key value had
a default value which was used not divided by 2, if the user was not
providing that key in the input file. This made my commit change the
code behaviour. In this commit I am doubling the key default value to
restore the older behaviour.