Fix still wrong number of ensemble in ROOT output
The number of ensemble was wrongly set in the ROOT output as it was
wrongly stored in the TTree object. In particular the RootOutput class
is first storing information in class members and then uses class
members to populate the ROOT TTree objects. However, the ensemble number
was stored in the class member only in the at_eventstart method. Now it
is also stored at_intermediate_time and at_eventend.