Fix bug in Experiment class about output clock increment
The output clock was incremented only before producing the intermediate
output was created. However, after the last intermediate output and
before the final output there was no increment of the output clock and
the final output function was using the time current time of the output
clock, which was wrongly set. Now the output clock is updated at the end
of the run_time_evolution function.

Note that some outputs were (apparently) not affected by this bug,
because they were writing the information about time exctracting it from
the particle metadata. However, the EventInfo handed over to the output
method at_eventend was still containing the wrong time value.