Suppress CMake CONFIGURE_LOG output level
This has been introduced in CMake 3.26 and I did not find a CMake
functionality to suppress it. By default CMake prints all of these
messages when e.g. using the try_compile() function and this blows up
output with very low-level messages. I therefore opted to suppress it
in our wrapper of the message() function.