Modify check about output quantities
These can now be also provided with Binary output and they must be
provided if the Binary format is used, because we want to make Binary
equivalent to ASCII. However, such should be ATM restricted to Particles
and Collisions content, only, and so has been implemented. Note that
then the quantities map is not checked anymore before been accessed with
the at method, since it is assumed to always contain the Particles or
Collisions list.