Change binary output classes constructors and philosophy
All binary output classes are now supposed to take a list of quantities
at construction time and will throw if such a list is passed
empty. Therefore it is client responsibility to use them as wished and
the function to create a binary output does exactly so. In this way it
is in particular easier to implement an alias for the OSCAR2013
quantities, since this has to be done before instantiating any binary
output object and the two responsibility will stay separated.