Fix potential bug in OutputParameters constructor
The quantities for particles and collisions outputs were taken
conditionally to the existence of the corresponding section in the
configuration file. However, the configuration file handed over to the
OutputParameters constructor had the 'Format' key from each output
content already extracted in Experiment. If such a key happened to be
the only existing one, then taking it in Experiment would have led to
the full content subsection be removed, which in turn made the if-clause
in the OutputParameter constructor be skipped, doing the wrong thing. In
that case the quantities map had missing entries, which in turn would
have triggered an exception being thrown when using std::map::at.