Fix bug about Binary output crashing at renaming time
If the Binary output was requested with the same quantitites of
Oscar2013_bin and also Oscar2013_bin was given as a format, SMASH was
failing because it tried to produce twice the same file and the
".unfinished-renaming mechansim" was not working. Since this usage would
make SMASH produce twice the same file, this has now been forbidden.