Improve Docker images and bump ROOT version
The ROOT version of the max image has been increased to 6.34 which is
not containing the memory leak existing in previous version which was
making SMASH test fail. Moreover, minor improvement have been done to
avoid Docker warnings. Software versions have been added into the max
image as label for later retrieval and the prompt setting both in Docker
and in Singularity has been moved to the min container and, doing so, it
should automatically be done in both images (the max image is built from
the min one).