Fix wrong default value of Monash tune key
Although set to false, which is the correct behavioural default for most
SMASH runs, the InputKey object's default value should have been
DefaultValue::Dependent as it is now. That is the way to signal in the
code that there is no a fixed default value and that this depends on
other aspects. Being this wrong, taking the key as introduced in commit
c258ff5  was resulting in not using the Monash tune even at high
energies, unless explicitly specified in the configuration (rarely the
case). Now that the default is marked as dependent, the key must be
taken explicitly specifying the default value and this is done as it was
before commit c258ff5.