Improve script to update Pythia allowing to use BSD sed
The script was assuming that the user was using GNU sed and, in
particular, sed was invoked using the -i option without a
suffix. Although this works with GNU sed, it does not with BSD sed, as
BSD implementation requires a suffix value to use sed inplace. The script
is now always giving a suffix, removing backup files if sed succeeds.